Make tweetbox and PostPage function components and changed them to typescript
