--- src/components/tweet-box/TweetBox.js
@@ -1,188 +1,0 @@
-import React, {useEffect, useState} from "react";
-import Button from "../button/Button";
-import TweetInput from "../tweet-input/TweetInput";
-import {useHttpRequestService} from "../../service/HttpRequestService";
-import {setLength, updateFeed} from "../../redux/user";
-import ImageContainer from "../tweet/tweet-image/ImageContainer";
-import {BackArrowIcon} from "../icon/Icon";
-import ImageInput from "../common/ImageInput";
-import {useTranslation} from "react-i18next";
-import {ButtonType} from "../button/StyledButton";
-import {StyledTweetBoxContainer} from "./TweetBoxContainer";
-import {StyledContainer} from "../common/Container";
-import {StyledButtonContainer} from "./ButtonContainer";
-import {useDispatch, useSelector} from "react-redux";
-
-const TweetBox = (props) => {
-    const {parentId, close, mobile, onCommentCreated} = props;
-    const [content, setContent] = useState("");
-    const [images, setImages] = useState([]);
-    const [imagesPreview, setImagesPreview] = useState([]);
-
-    const {length, query} = useSelector((state) => state.user);
-    const httpService = useHttpRequestService();
-    const dispatch = useDispatch();
-    const {t} = useTranslation();
-    const service = useHttpRequestService()
-    const [user, setUser] = useState()
-
-
-    useEffect(() => {
-        handleGetUser().then(r => setUser(r))
-    }, []);
-
-    const handleGetUser = async () => {
-        return await service.me()
-    }
-
-    const handleChange = (e) => {
-        setContent(e.target.value);
-    };
-    const handleSubmit = async () => {
-        try {
-            setContent("");
-            setImages([]);
-            setImagesPreview([]);
-            
-            let result;
-            if (parentId) {
-                // This is a comment - handle image uploads for comments
-                let imageKeys = [];
-                
-                if (images && images.length > 0) {
-                    const imageRequests = images.map((file) => {
-                        return {
-                            fileExtension: file.name ? file.name.split('.').pop() : 'jpg',
-                            contentType: file.type || 'image/jpeg'
-                        };
-                    });
-
-                    const uploadUrlResponse = await httpService.getImageUploadUrls('comment/images/upload-urls', { images: imageRequests });
-
-                    if (uploadUrlResponse && uploadUrlResponse.uploads) {
-                        // Upload each image to S3 using presigned URLs
-                        for (let i = 0; i < images.length; i++) {
-                            const file = images[i];
-                            const uploadInfo = uploadUrlResponse.uploads[i];
-
-                            if (!uploadInfo || !uploadInfo.uploadUrl || !uploadInfo.imageKey) {
-                                throw new Error(`Invalid upload data for image ${i + 1}`);
-                            }
-
-                            const { uploadUrl, imageKey } = uploadInfo;
-
-                            const uploadResponse = await fetch(uploadUrl, {
-                                method: 'PUT',
-                                body: file,
-                            });
-
-                            if (!uploadResponse.ok) {
-                                throw new Error(`Failed to upload image ${i + 1}`);
-                            }
-
-                            imageKeys.push(imageKey);
-                        }
-                    }
-                }
-                
-                // Create comment with images
-                const commentData = {
-                    content: content,
-                    images: imageKeys
-                };
-                result = await httpService.createComment(parentId, commentData);
-            } else {
-                // This is a regular post, use the existing post creation logic
-                const postData = {
-                    content: content,
-                    images: images, // Pass File objects for posts
-                    parentId: parentId
-                };
-                result = await httpService.createPost(postData);
-            }
-            
-            // If this is a comment (has parentId), call the callback instead of updating global feed
-            if (parentId && onCommentCreated) {
-                onCommentCreated();
-            } else {
-                // Only update global feed if this is a regular post (no parentId)
-                dispatch(setLength(length + 1));
-                const posts = await httpService.getPosts(query);
-                dispatch(updateFeed(posts));
-            }
-            
-            close && close();
-        } catch (e) {
-            console.log(e);
-        }
-    };
-
-    const handleRemoveImage = (index) => {
-        const newImages = images.filter((i, idx) => idx !== index);
-        const newImagesPreview = newImages.map((i) => URL.createObjectURL(i));
-        setImages(newImages);
-        setImagesPreview(newImagesPreview);
-    };
-
-    const handleAddImage = (newImages) => {
-        setImages(newImages);
-        const newImagesPreview = newImages.map((i) => URL.createObjectURL(i));
-        setImagesPreview(newImagesPreview);
-    };
-
-    return (
-        <StyledTweetBoxContainer>
-            {mobile && (
-                <StyledContainer
-                    flexDirection={"row"}
-                    justifyContent={"space-between"}
-                    alignItems={"center"}
-                >
-                    <BackArrowIcon onClick={close}/>
-                    <Button
-                        text={"Tweet"}
-                        buttonType={ButtonType.DEFAULT}
-                        size={"SMALL"}
-                        onClick={handleSubmit}
-                        disabled={content.length === 0}
-                    />
-                </StyledContainer>
-            )}
-            <StyledContainer style={{width: "100%"}}>
-                <TweetInput
-                    onChange={handleChange}
-                    maxLength={240}
-                    placeholder={t("placeholder.tweet")}
-                    value={content}
-                    src={user?.profilePicture}
-                />
-                <StyledContainer padding={"0 0 0 10%"}>
-                    <ImageContainer
-                        editable
-                        images={imagesPreview}
-                        removeFunction={handleRemoveImage}
-                    />
-                </StyledContainer>
-                <StyledButtonContainer>
-                    <ImageInput setImages={handleAddImage} parentId={parentId}/>
-                    {!mobile && (
-                        <Button
-                            text={"Tweet"}
-                            buttonType={ButtonType.DEFAULT}
-                            size={"SMALL"}
-                            onClick={handleSubmit}
-                            disabled={
-                                content.length <= 0 ||
-                                content.length > 240 ||
-                                images.length > 4 ||
-                                images.length < 0
-                            }
-                        />
-                    )}
-                </StyledButtonContainer>
-            </StyledContainer>
-        </StyledTweetBoxContainer>
-    );
-};
-
-export default TweetBox;
