+++ src/pages/post-page/PostPage.tsx
@@ -8,18 +8,19 @@ import TweetBox from "../../components/tweet-box/TweetBox";
 import { StyledH5 } from "../../components/common/text";
 import { StyledFeedContainer } from "../home-page/components/contentContainer/FeedContainer";
 import CommentFeed from "../../components/feed/CommentFeed";
+import { Post } from "../../service";
 
 const PostPage = () => {
-  const { id: postId } = useParams();
-  const [post, setPost] = useState(undefined);
-  const [refreshComments, setRefreshComments] = useState(0);
+  const { id: postId } = useParams<{ id: string }>();
+  const [post, setPost] = useState<Post | undefined>(undefined);
+  const [refreshComments, setRefreshComments] = useState<number>(0);
   const service = useHttpRequestService();
 
-  const fetchPost = async () => {
+  const fetchPost = async (): Promise<void> => {
     if (!postId) return;
 
     try {
-      const res = await service.getPostById(postId);
+      const res: Post = await service.getPostById(postId);
       setPost(res);
     } catch (e) {
       console.log(e);
@@ -30,9 +31,9 @@ const PostPage = () => {
     fetchPost();
   }, [postId]); // This will re-run when postId changes
 
-  const handleCommentCreated = () => {
+  const handleCommentCreated = (): void => {
     // Trigger comment refresh by updating the refreshComments counter
-    setRefreshComments((prev) => prev + 1);
+    setRefreshComments((prev: number) => prev + 1);
   };
 
   return (
